Max of 32-bit values can be processed

+++ main.h
@@ -15,7 +15,7 @@ int Putchar(char ch);
 int Puts(char *string);
 int _printf(const char *format, ...);
 void DecToHex(char *outBuffer, uint64_t number);
-int DecimalToBinary(uint64_t decimalNum);
+int DecimalToBinary(uint32_t decimalNum);
 int PutsCustom(char *string);
 int PrintfFlags(char flag, va_list arguments);
 int PutsReverse(char *string);
